Refactored the EasyAdminExtension DI extension
